backend/requirements.txt
Flask==3.0.3
flask-restx==1.3.0
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
Flask-Limiter==3.8.0
SQLAlchemy==2.0.34
email-validator==2.2.0
Werkzeug==3.0.4
python-dotenv==1.0.1
passlib==1.7.4
itsdangerous==2.2.0
blinker==1.8.2
flask_limiter
